Add files via upload
"routh_proj.mlx" is the main Matlab program that takes the input transfer function from user and extracts the coefficient polynomials of the numerator and denominator and calls "RouthCriteria.m" class which caries the logic of Routh stability. You can expect output to be the Routh table, command including whether the system is stable or not with the number of positive poles and the root locus of the system. 


Also, this code determines stability even if there is a special case.